requirements.txt
accelerate==1.3.0
av==14.3.0
diffusers==0.32.0
einops==0.8.1
face_alignment==1.4.1
huggingface_hub==0.29.1
imageio==2.37.0
insightface==0.7.3
lpips==0.1.4
matplotlib==3.10.1
numpy==2.0.1
omegaconf==2.3.0
onnxruntime==1.21.1
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pandas==2.2.3
Pillow==11.2.1
PyYAML==6.0.2
scipy==1.15.2
scikit-image==0.25.2
torch==2.5.1
torch-fidelity==0.3.0
torchmetrics==1.7.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.49.0